Fix: 그룹 생성 시 알람 음량 default 값 설정

+++ data/src/main/java/com/asap/data/remote/datasource/GroupRemoteDataSource.kt
@@ -96,7 +96,7 @@ class GroupRemoteDataSource @Inject constructor(
         alarmTime: String,
         alarmType: String,
         alarmUnlockContents: String,
-        alarmVolume: Int?,
+        alarmVolume: Float?,
         description: String,
         deviceToken: String,
         deviceType: String,
